Add FluentImportManager, the fluent counterpart of ImportManagerExtension
FluentJavaClassGenerator#buildClass had been made public so
FunctionGeneratorHelper could call it directly to assemble a single
function's class file from its CodeRenderer. That widened an internal
detail of the generator base class purely to serve one external caller,
instead of exposing an independently-usable service the way
ImportManagerExtension always was for the legacy Xtend path.

Extract buildClass into FluentImportManager, injected the same way
XtendJavaClassGenerator injects ImportManagerExtension, and have
FunctionGeneratorHelper depend on it directly instead of reaching
through the generator instance. buildClass reverts to protected.
